[SubtasksJS-6] Вынесла обработку результата валидации, установку текста в инпут в отдельные функции, убрала лишний класс на форме
