+++ SubtasksJs/Nika/SubtaskJS-6/src/js/setTextForInputFile.js
@@ -1,0 +1,14 @@
+export function setTextForInputFile(
+  file,
+  inputFileTextID,
+  initialTextforInputFile
+) {
+  const inputFileText = document.getElementById(inputFileTextID);
+  if (file) {
+    inputFileText.innerText = file.name;
+    return true;
+  } else {
+    inputFileText.innerText = initialTextforInputFile;
+    return false;
+  }
+}
