+++ SubtasksJs/Nika/SubtaskJS-6/src/styles/styles.scss
@@ -29,16 +29,12 @@ $color-dark-brown: #721b1b;
   justify-content: center;
   flex-wrap: wrap;
   gap: rem(20);
-  visibility: hidden;
   width: 100%;
   max-width: rem(600);
   margin: 0 auto rem(20);
   padding: rem(20);
   border: 2px solid $color-grey;
   border-radius: rem(10);
-  &_visible {
-    visibility: visible;
-  }
   &__title {
     width: 100%;
     font-size: rem(22);
@@ -92,7 +88,8 @@ $color-dark-brown: #721b1b;
     color: $color-black;
     border-radius: rem(4);
     border: 1px solid $color-black;
-    &:invalid, &_error {
+    &:invalid,
+    &_error {
       border-color: $color-red;
     }
     &_file {
@@ -153,11 +150,6 @@ $color-dark-brown: #721b1b;
     text-align: center;
     border: 2px solid $color-red;
     border-radius: rem(10);
-  }
-  &__error-text {
-    color: $color-red;
-    font-size: rem(18);
-    width: 100%;
   }
   &__control {
     display: flex;
