+++ SubtasksJs/Nika/SubtaskJS-6/src/template.html
@@ -8,7 +8,7 @@
   <body>
     <main id="main">
       <h1 class="title">Генератор JSON</h1>
-      <form class="form form_visible" id="generator">
+      <form class="form" id="generator">
         <h2 class="form__title">Сгенерировать форму</h2>
         <label class="form__label form__label_file" for="selectFile">
           <span class="form__file-text" id="selectFileText">Выбрать файл</span>
